maybe adding asserts to header_map.h
