+++ include/envoy/server/BUILD
@@ -25,6 +25,7 @@ envoy_cc_library(
         ":config_tracker_interface",
         "//include/envoy/buffer:buffer_interface",
         "//include/envoy/http:codes_interface",
+        "//include/envoy/http:header_map_interface",
         "//include/envoy/network:listen_socket_interface",
     ],
 )
