+++ source/common/common/logger.h
@@ -5,8 +5,6 @@
 #include <string>
 #include <vector>
 
-#include "envoy/access_log/access_log.h"
-#include "envoy/filesystem/filesystem.h"
 #include "envoy/thread/thread.h"
 
 #include "common/common/fmt.h"
@@ -106,22 +104,6 @@ protected:
 private:
   SinkDelegate* previous_delegate_;
   DelegatingLogSinkPtr log_sink_;
-};
-
-/**
- * SinkDelegate that writes log messages to a file.
- */
-class FileSinkDelegate : public SinkDelegate {
-public:
-  FileSinkDelegate(const std::string& log_path, AccessLog::AccessLogManager& log_manager,
-                   DelegatingLogSinkPtr log_sink);
-
-  // SinkDelegate
-  void log(absl::string_view msg) override;
-  void flush() override;
-
-private:
-  Filesystem::FileSharedPtr log_file_;
 };
 
 /**
